Migrate core components to TypeScript
- Converted Home.jsx → Home.tsx with HomeProps interface
- Converted MainLayout.jsx → MainLayout.tsx with MainLayoutProps interface
- Converted Navbar.jsx → Navbar.tsx with NavbarProps interface
- Converted App.jsx → App.tsx as React.FC
- Added NavLinkClassProps interface for navbar link styling
- Updated imports to use TypeScript components
- Verified application compiles successfully with TypeScript

Next: Migrate remaining components (HomePage, AstroGallery, etc.)
